# SiteManager/public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: Site Manager API
  description: Site Manager API Documentation
  contact:
    email: ndungudennis250@gmail.com
  version: 1.0.0
servers:
  - url: http://172.105.90.112/site-manager-backend/SiteManager
    description: Production server
  - url: http://localhost
    description: Development server
paths:
...
  /api/clockIn:
    post:
      tags:
        - ClockIns
      summary: Clock in
      description: Clock in
      operationId: clockIn
      requestBody:
        description: Pass clock in details
        required: true
        content:
          application/json:
            schema:
              required:
                - siteManagerId
                - projectId
                - workerId
                - clockInTime
              properties:
                siteManagerId:
                  type: integer
                  format: int64
                  example: 1
                projectId:
                  type: integer
                  format: int64
                  example: 1
                workerId:
                  type: integer
                  format: int64
                  example: 1
                clockInTime:
                  type: string
                  format: date-time
                  example: '2023-07-20'
              type: object
      responses:
        '201':
          description: Clocked in successfully
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Clocked in successfully
                type: object
        '404':
          description: Worker does not exist
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: Worker does not exist
                type: object
        '500':
          description: Server error
...
  /api/report/{projectId}:
    get:
      tags:
        - Reports
      summary: Generate report
      description: Generate report
      operationId: generateReport
      parameters:
        - name: projectId
          in: path
          description: The project id
          required: true
          schema:
            type: integer
            format: int64
            example: 1
        - name: startDate
          in: query
          description: The start date
          required: false
          schema:
            type: string
            format: date-time
            example: '2023-07-20'
        - name: endDate
          in: query
          description: The end date
          required: false
          schema:
            type: string
            format: date-time
            example: '2023-07-20'
      responses:
        '200':
          description: Report generated successfully
        '404':
          description: No report found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: No report found
                type: object
        '500':
          description: Server error
  /api/workerReport/{workerId}:
    get:
      tags:
        - Reports
      summary: Generate worker report
      description: Generate worker report
      operationId: generateWorkerReport
      parameters:
        - name: workerId
          in: path
          description: The worker id
          required: true
          schema:
            type: integer
            format: int64
            example: 1
        - name: startDate
          in: query
          description: The start date
          required: false
          schema:
            type: string
            format: date-time
            example: '2023-07-20'
        - name: endDate
          in: query
          description: The end date
          required: false
          schema:
            type: string
            format: date-time
            example: '2023-07-20'
      responses:
        '200':
          description: Report generated successfully
        '404':
          description: No report found
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: No report found
                type: object
        '500':
          description: Server error
...